+++ src/lib/random.c
@@ -14,6 +14,6 @@ tinymt32_t tinymt;
 //generates random number in range (inclusive lower and upper bounds)
 long randrange(long lower, long upper) {
     uint32_t r = tinymt32_generate_uint32(&tinymt);
-    unsigned rand = lower + r / (UINT32_MAX / (upper - lower + 1) + 1);
+    long rand = lower + r / (UINT32_MAX / (upper - lower + 1) + 1);
     return rand;
 }
