+++ src/web_server.c
@@ -607,7 +607,7 @@ static void ev_handler_redirect(struct mg_connection *nc, int ev, void *ev_data,
                 break;
             }
 
-            sds host_header = sdscatlen(sdsempty(), host_hdr->ptr, (int)host_hdr->len);
+            sds host_header = sdscatlen(sdsempty(), host_hdr->ptr, host_hdr->len);
             int count = 0;
             sds *tokens = sdssplitlen(host_header, (ssize_t)sdslen(host_header), ":", 1, &count);
             sds s_redirect = sdscatfmt(sdsempty(), "https://%S", tokens[0]);
